% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang_pga_rotd50_calc.R
\name{ang_pga_rotd50_calc}
\alias{ang_pga_rotd50_calc}
\title{Rotation angle of PGA RotD50 Calculation}
\usage{
ang_pga_rotd50_calc(h1, h2)
}
\arguments{
\item{h1}{A vector of time series for the 1st horizontal component}

\item{h2}{A vector of time series for the 2nd horizontal component}
}
\value{
The rotation angle for PGA RotD50
}
\description{
This function calculates the rotation angle (relative to the azimuth of h1) of PGA RotD50
}
