% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_pass.R
\name{bw_pass}
\alias{bw_pass}
\title{Low-pass Butterworth filter}
\usage{
bw_pass(ts, dt, fc, nPole = 4, is_causal = FALSE, order_zero_padding = 0)
}
\arguments{
\item{ts}{An array of time series}

\item{dt}{The time step}

\item{fc}{The corner frequency for low-pass/high-pass}

\item{nPole}{The pole parameter for the order, where the order is 2 * nPole. Positive for low-pass; Negative for high-pass
Generally, nPole = 4 for low-pass and nPole = -5 for high-pass}

\item{is_causal}{A binary to indicator if causal is applied. Apply causal if is_causal = TRUE and apply acausal if is_causal = FALSE}

\item{order_zero_padding}{The order needs to be added for zeroes padding at the end of recordings to
increase the number of data points to a power of 2 (Kishida T., Darragh R.B., Chiou B.S.J., Bozorgnia Y., Mazzoni S.,
Contreras V., Boroschek R., Rojas F., Stewart J.P., 2020.
Chapter 3: Ground Motions and Intensity Measures, in Data Resources for NGA-Subduction Project,
PEER Report 2020/02, J.P. Stewart (editor), Pacific Earthquake Engineering Research Center, UC Berkeley (headquarters).). Default is 0 (no zero-padding).}
}
\value{
The filtered time series
}
\description{
This function applies low-pass Butterworth filter to process time series (for both causal and acausal)
}
