% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas.R
\name{fas_cal}
\alias{fas_cal}
\title{A function for Fourier Amplitude Spectrum (FAS) calculation}
\usage{
fas_cal(ts, dt, max_order = 0)
}
\arguments{
\item{ts}{An array of time series}

\item{dt}{The time step}

\item{max_order}{An integer, the maximum factor after padding 0 at the end;
if max_order = 0, then no zero padding;
if max_order = 3, the length of ts becomes at least 4 times.}
}
\value{
A list is returned with FFT of ts, FAS, the associated frequency of FAS, the associated phase
}
\description{
This function computes FAS (with and without 0 padding)
}
\keyword{FAS}
