% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_plt_select.R
\name{fd_plt_select}
\alias{fd_plt_select}
\title{Frequency-domain plot and window selection}
\usage{
fd_plt_select(
  hvsr_list,
  freq_hv_mean,
  freq_min,
  freq_max,
  hpass_fc,
  lpass_fc,
  distribution
)
}
\arguments{
\item{hvsr_list}{A list. Each element is a window HVSR}

\item{freq_hv_mean}{The frequency}

\item{freq_min}{The minimum frequency of interest}

\item{freq_max}{The maximum frequency of interest}

\item{hpass_fc}{High-pass corner frequency applied in filter}

\item{lpass_fc}{Low-pass corner frequency applied in filter}

\item{distribution}{The distribution assumption on the HVSR amplitudes. It can take "normal" or "log_normal".}
}
\value{
The lefted indices of windows after frequency-domain rejection
}
\description{
This function plots each window in frequency domain and let users to manually select windows for rejection
}
