% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_fit.R
\name{gaussian_gen_fit}
\alias{gaussian_gen_fit}
\title{Generalized Gaussian pulse function}
\usage{
gaussian_gen_fit(par, x_freq)
}
\arguments{
\item{par}{The array of fitted Gaussian pulse function parameters, c0, c1, fp, and w}

\item{x_freq}{An array of frequency where we predict HVSR}
}
\description{
This is a generalized Gaussian pulse function used to fit HVSR peaks
}
