% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_proc.R
\name{hv_proc}
\alias{hv_proc}
\title{HVSR-processing main function}
\usage{
hv_proc(
  is_noise = TRUE,
  h1,
  h2,
  v,
  dt,
  eqk_filepath,
  output_dir,
  output_pf_flnm = "Test_",
  distribution = "normal",
  robust_est = FALSE,
  pre_filter_flag = FALSE,
  pre_filter_is_causal = FALSE,
  pre_filter_hpass_fc = 0.1,
  pre_filter_lpass_fc = NA,
  pre_filter_nPole_hp = 5,
  pre_filter_nPole_lp = 4,
  pre_filter_order_zero_padding = 2,
  pre_filter_t_front = 5,
  pre_filter_t_end = 5,
  filter_flag = TRUE,
  is_causal = FALSE,
  hpass_fc = 0.1,
  lpass_fc = NA,
  nPole_hp = 5,
  nPole_lp = 4,
  order_zero_padding = 2,
  detrend = 1,
  taper_flag = TRUE,
  t_front = 5,
  t_end = 5,
  horizontal_comb = "geometric_mean",
  ko_smooth_flag = TRUE,
  ko_smooth_b = 20,
  parzen_flag = FALSE,
  parzen_bwidth = 1.5,
  win_width = 150,
  overlapping = 0,
  sta_lta_flag = FALSE,
  sta_lta_moving_term_len = 1,
  short_term_len = 1,
  long_term_len = 30,
  polar_curves_flag = TRUE,
  deg_increment = 10,
  resample_lin2log = TRUE,
  deci_mean_factor = 10,
  deci_polar_factor = 10,
  output_freq_min = 0.01,
  output_freq_max = 50,
  output_selected_ts = FALSE,
  output_removed_ts = FALSE,
  output_selected_hvsr = TRUE,
  output_removed_hvsr = FALSE,
  output_mean_curve = TRUE,
  output_polar_curves = TRUE,
  output_metadata = TRUE
)
}
\arguments{
\item{is_noise}{A binary indicating if the input data is ambient noise. is_noise = TRUE if input is noise;
otherwise, if is_noise = FALSE, the input is corrected strong motion. Note when is_noise = TRUE, h1,
h2, v, and, dt are needed; when is_noise = FALSE, eqk_filepath is needed.}

\item{h1}{The first component of horizontal time series}

\item{h2}{The second component of horizontal time series}

\item{v}{The vertical component of time series}

\item{dt}{The time step}

\item{eqk_filepath}{The file path where it saves the file of corrected strong motions data in standard format}

\item{output_dir}{The directory of output results}

\item{output_pf_flnm}{The prefix name of output file names}

\item{distribution}{The distribution assumption on the HVSR amplitudes. It can take "normal" or "log_normal".
We recommend normal for noise data and log_normal for corrected strong motions}

\item{pre_filter_flag}{The flag indicates if a pre-filter for the entire time series is applied (before windowing).}

\item{pre_filter_is_causal}{A binary to indicator if causal is applied in pre-filter}

\item{pre_filter_hpass_fc}{The corner frequency for high-pass in per-filter}

\item{pre_filter_lpass_fc}{The corner frequency for low-pass in per-filter}

\item{pre_filter_nPole_hp}{The pole parameter for high-pass in per-filter. Default is 5}

\item{pre_filter_nPole_lp}{The pole parameter for low-pass in per-filter. Default is 4}

\item{pre_filter_order_zero_padding}{The order needs in per-filter to be added for zeroes padding at the end of recordings to
increase the number of data points to a power of 2. Default is 2.}

\item{pre_filter_t_front}{The percentage of taperring for the beginning of the time series in the pre-filter.}

\item{pre_filter_t_end}{The percentage of taperring for the end of the time series in the pre-filter.}

\item{filter_flag}{The flag indicates if filter is applied to noise. The corrected strong motions do not apply}

\item{is_causal}{A binary to indicator if causal is applied to noise. Apply causal if is_causal = TRUE and apply acausal if is_causal = FALSE}

\item{hpass_fc}{The corner frequency for high-pass filter}

\item{lpass_fc}{The corner frequency for low-pass filter}

\item{nPole_hp}{The pole parameter for high-pass. Default is 5}

\item{nPole_lp}{The pole parameter for low-pass. Default is 4}

\item{order_zero_padding}{The order needs to be added for zeroes padding at the end of recordings to
increase the number of data points to a power of 2. Default is 2}

\item{detrend}{The indicator specifies which detrend method is used. 0: no detrend; 1: mean removal; 2: linear trend removal; 6: fifth order polynomial detrend.}

\item{taper_flag}{The flag indicates if Taper is applied to noise. The corrected strong motions do not apply}

\item{t_front}{The percentage of taperring for the beginning of the time series.}

\item{t_end}{The percentage of taperring for the end of the time series.}

\item{horizontal_comb}{The parameter specifies the combination of two horizontal components. ps_RotD50: rotated combination at the angle where PGA is median; geometric_mean: geometric mean (sqrt(h1(f) * h2(f))); squared_average: squared average (sqrt((h1(f)^2 + h2(f)^2)/2))}

\item{ko_smooth_flag}{The flag indicates if KO smoothing is applied.}

\item{ko_smooth_b}{The coefficient of bandwidth. Default is 20. A smaller value will lead to more smoothing}

\item{parzen_flag}{The flag indicates if Parzen smoothing is applied. Default is FALSE since KO smoothing is applied by default}

\item{parzen_bwidth}{The bandwidth. A larger value will lead to more smoothing}

\item{win_width}{The window length for noise data. The unit is second. The win_width equals to one earthquake duration if
the input is corrected strong motion}

\item{overlapping}{Overlapping between two windows, unit is second. Default is 0 without any overlapping}

\item{sta_lta_flag}{The flag indicates if STA/LTA is calculated}

\item{sta_lta_moving_term_len}{The length of moving step for STA/LTA, in second}

\item{short_term_len}{The short term length for STA, in second}

\item{long_term_len}{The long term length for LTA, in second}

\item{polar_curves_flag}{The flag indicates if polar curves are calculated}

\item{deg_increment}{The degree increment for HVSR polar curves}

\item{resample_lin2log}{The flag indicates if resampling frequency from equal-spaced linear scale to equal-spaced log scale}

\item{deci_mean_factor}{An integer for decimation factor for HVSR mean curve}

\item{deci_polar_factor}{An integer for decimation factor for HVSR polar curves}

\item{output_freq_min}{The minimum output frequency of HVSR}

\item{output_freq_max}{The maximum output frequency of HVSR}

\item{output_selected_ts}{The flag indicates if output selected time seriese windows}

\item{output_removed_ts}{The flag indicates if output rejected time seriese windows}

\item{output_selected_hvsr}{The flag indicates if output selected HVSR curves}

\item{output_removed_hvsr}{The flag indicates if output rejected HVSR curves}

\item{output_mean_curve}{The flag indicates if output HVSR mean curve}

\item{output_polar_curves}{The flag indicates if output HVSR polar curves}

\item{output_metadata}{The flag indicates if output HVSR processing meta data}
}
\value{
The results are saved and written in files in the specified output_dir
}
\description{
This is the main function for HVSR processing (ambient-noise or corrected strong motions)
}
