% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvsrProc.R
\docType{package}
\name{hvsrProc-package}
\alias{hvsrProc-package}
\alias{hvsrProc}
\title{HVSR Processing Package}
\description{
This package is developed for processing and calculating
HVSR from noise-based and earthquake-based time series data.
}
\note{
There are 17 hvsrProc functions:
\code{ang_pga_rotd50_calc, bw_pass, bw_resp, fas_cal, fd_plt_select,
hv_proc, hvsr_win_calc, ko_smooth, log10_ticks, logspace,
mean_removal, parzen_smooth, plot_log10, pre_proc,
sta_lta_calc, taper, td_plt_select}

Of which \code{hv_proc} is main function.
}
\examples{
## Not run:
##   See examples in the help files for all functions.
## End(Not run)

}
\author{
Pengfei Wang
}
