% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_log10.R
\name{log10_ticks}
\alias{log10_ticks}
\title{A function to add log10 ticks}
\usage{
log10_ticks(x, y, log10_scale = "x", tick_type = "exp", minor_tick = TRUE)
}
\arguments{
\item{x}{The array of x values}

\item{y}{The array of y values}

\item{log10_scale}{A string takes "x", "y", or "xy". It indicates the axix that will be plotted in log10 scale}

\item{tick_type}{A string takes "exp" (the ticks are expressed by 10^) or "lin" (the ticks are expressed by 100..)}

\item{minor_tick}{Binary, indicates if minor ticks are plotted}
}
\description{
This function generates log10 ticks for the figure in x-axis / y-axis
}
