% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logspace.R
\name{logspace}
\alias{logspace}
\title{Generate linear log space}
\usage{
logspace(x1, x2, log_type = "log10", n = 50)
}
\arguments{
\item{x1}{The start point}

\item{x2}{The end point}

\item{log_type}{The type of log scale. "log10" is for log10(), "ln" is for natural log()}

\item{n}{The number of points to be generated}
}
\value{
An array of n points with equal space in log scale
}
\description{
This function generates n points with equal space in log scale
}
