% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parzen_smooth}
\alias{parzen_smooth}
\title{Parzen smoothing}
\usage{
parzen_smooth(freq, amp, b = 1.5)
}
\arguments{
\item{freq}{An array of the frequency}

\item{amp}{An array of the corresponding FAS}

\item{b}{The bandwidth. A larger value will lead to more smoothing}
}
\value{
The smoothed FAS
}
\description{
Apply Parzen smoothing to the Fourier Amplitude Spectra (FAS). The key formula is referred from Eq (5) on the page: http://www2.kobe-u.ac.jp/~nagaotak/poster2017wcee.pdf
}
