% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_fit.R
\name{peak_fit_auto}
\alias{peak_fit_auto}
\title{A function for HVSR peak detection and fitting automatically}
\usage{
peak_fit_auto(
  x_freq,
  hvsr_mean,
  hvsr_sd,
  ratio_thres = 0.7,
  amp_thres = 1.5,
  k_cv_thres = 1,
  max_freq = 15,
  min_freq = 0.1,
  cp = 0.005,
  x_jump = 0.35,
  peak_width_thres_L = 2.3
)
}
\arguments{
\item{x_freq}{The array of frequnecy (in Hz)}

\item{hvsr_mean}{The array of HVSR mean curve}

\item{hvsr_sd}{The array of HVSR amplitude standard deviation}

\item{ratio_thres}{The ratio threshold to determine if a peak is "clear" enough}

\item{amp_thres}{The minimum required peak amplitude for a clear peak}

\item{k_cv_thres}{The threshold of coefficient of variation for a clear peak}

\item{max_freq}{The maximum frequency to be considered as a peak}

\item{min_freq}{The minimum frequency to be considered as a peak}

\item{cp}{The complexity parameter for regression tree. A smaller number leads to a finer fitting.}

\item{x_jump}{The step width to determine a stable plateau}

\item{peak_width_thres_L}{The threshold of peak width (in natural log of frequency) for a clear peak}
}
\description{
This function automatically detect if a clear peak exits. If there is a clear peak,
a Gaussian pulse function is used to fit the peak. The fitted parameters are returned.
}
