% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_fit.R
\name{peak_fit_manual}
\alias{peak_fit_manual}
\title{A function for HVSR peak detection and fitting manually}
\usage{
peak_fit_manual(x_freq, hvsr_mean, fit_range_lb, fit_range_ub)
}
\arguments{
\item{x_freq}{The array of frequnecy (in Hz)}

\item{hvsr_mean}{The array of HVSR mean curve}

\item{fit_range_lb}{The lower frequency bound for a HVSR peak (in Hz)}

\item{fit_range_ub}{The upper frequency bound for a HVSR peak (in Hz)}
}
\description{
This function takes the user defined frequency range (lower bound and upper bound) and
use the Gaussian pulse function to fit the peak. The fitted parameters are returned.
}
