% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log10.R
\name{plot_log10}
\alias{plot_log10}
\title{A function to plot a figure in log10 scale}
\usage{
plot_log10(
  x,
  y,
  log10_scale = "x",
  tick_type = "lin",
  minor_tick = TRUE,
  xlab = NULL,
  ylab = NULL,
  type = NULL,
  main = NULL
)
}
\arguments{
\item{x}{The array of x values}

\item{y}{The array of y values}

\item{log10_scale}{A string takes "x", "y", or "xy". It indicates the axix that will be plotted in log10 scale}

\item{tick_type}{A string takes "exp" (the ticks are expressed by 10^) or "lin" (the ticks are expressed by 100..)}

\item{minor_tick}{Binary, indicates if minor ticks are plotted}

\item{xlab}{A title for the x axis}

\item{ylab}{A title for the y axis}

\item{type}{What type of plot should be drawn. The setting is the same as plot()}

\item{main}{An overall title for the plot}
}
\description{
This function generates a figure in log10 scale
}
