% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_plot.R
\name{polar_plot}
\alias{polar_plot}
\title{A function to generate polar plot}
\usage{
polar_plot(
  x,
  y,
  z,
  log10_scale = "x",
  tick_type = "lin",
  minor_tick = TRUE,
  xlab = "Freq (Hz)",
  ylab = "Azimuth (deg)"
)
}
\arguments{
\item{x}{The array of x values, e.g., frequency array}

\item{y}{The array of y values, e.g., azimuth angles}

\item{z}{The numeric matrix containing the values to be plotted. It must have the dimension of len(x) by len(y).}

\item{log10_scale}{A string takes "x", "y", or "xy". It indicates the axix that will be plotted in log10 scale}

\item{tick_type}{A string takes "exp" (the ticks are expressed by 10^) or "lin" (the ticks are expressed by 100..)}

\item{minor_tick}{Binary, indicates if minor ticks are plotted}

\item{xlab}{A title for the x axis}

\item{ylab}{A title for the y axis}
}
\description{
This function generates a polar plot figure (heatmap)
}
