% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.R
\name{pre_proc}
\alias{pre_proc}
\title{Pre-processing function}
\usage{
pre_proc(
  ts,
  dt,
  dc_flag = TRUE,
  taper_flag = TRUE,
  t_front,
  t_end,
  filter_flag = TRUE,
  fc,
  nPole,
  is_causal,
  order_zero_padding
)
}
\arguments{
\item{ts}{An array of time series}

\item{dt}{The time step}

\item{dc_flag}{The flag indicates if mean removal is applied}

\item{taper_flag}{The flag indicates if Taper is applied}

\item{t_front}{A number, the percentage of taperring for the beginning of the time series}

\item{t_end}{A number, the percentage of taperring for the end of the time series}

\item{filter_flag}{The flag indicates if filter is applied}

\item{fc}{The corner frequency for low-pass/high-pass}

\item{nPole}{The parameter for the order, where the order is 2 * nPole. Positive for low-pass; Negative for high-pass
Generally, nPole = 4 for low-pass and nPole = -5 for high-pass}

\item{is_causal}{A binary to indicator if causal is applied. Apply causal if is_causal = TRUE and apply acausal if is_causal = FALSE}

\item{order_zero_padding}{The order needs to be added for zeroes padding at the end of recordings to
increase the number of data points to a power of 2 (Kishida T., Darragh R.B., Chiou B.S.J., Bozorgnia Y., Mazzoni S.,
Contreras V., Boroschek R., Rojas F., Stewart J.P., 2020.
Chapter 3: Ground Motions and Intensity Measures, in Data Resources for NGA-Subduction Project,
PEER Report 2020/02, J.P. Stewart (editor), Pacific Earthquake Engineering Research Center, UC Berkeley (headquarters).)}
}
\value{
The filtered time series
}
\description{
This function pre-process time series (noise). The pre-processing includes mean removal, Taper, and Butterworth filter
}
