% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sta_lta_calc}
\alias{sta_lta_calc}
\title{STA to LTA calculation}
\usage{
sta_lta_calc(ts, short_term, long_term, moving_term)
}
\arguments{
\item{ts}{An array of time series}

\item{short_term}{An integer, short term length for STA (number of points)}

\item{long_term}{An integer, long term length for LTA (number of points)}

\item{moving_term}{An integer, moving step of the preceding point (number of points)}
}
\value{
A vector, the STA/LTA ratios
}
\description{
Calculate the ratio of Short Term Average (STA) to Long Term Average (LTA)
}
