% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{taper}
\alias{taper}
\title{Taper function}
\usage{
taper(ts, t_front = 5, t_end = 5)
}
\arguments{
\item{ts}{An array of time series}

\item{t_front}{A number, the percentage of taperring for the beginning of the time series}

\item{t_end}{A number, the percentage of taperring for the end of the time series}
}
\value{
The Taperred time series.
}
\description{
Apply Taper to the time series
}
