% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_plt_select.R
\name{td_plt_select}
\alias{td_plt_select}
\title{Time-domain plot and window selection}
\usage{
td_plt_select(
  h1_wins,
  h2_wins,
  v_wins,
  dt,
  sta_lta_flag,
  h1_stalta,
  h2_stalta,
  v_stalta,
  visual_deci_flag = TRUE,
  visual_deci = NA
)
}
\arguments{
\item{h1_wins}{A list. Each element is a window time series for 1st horizontal component}

\item{h2_wins}{A list. Each element is a window time series for 2nd horizontal component}

\item{v_wins}{A list. Each element is a window time series for vertical component}

\item{dt}{The time step}

\item{sta_lta_flag}{The flag indicates if STA/LTA is calculated}

\item{h1_stalta}{A list. Each element is a STA/LTA for a window of 1st horizontal component}

\item{h2_stalta}{A list. Each element is a STA/LTA for a window of 2nd horizontal component}

\item{v_stalta}{A list. Each element is a STA/LTA for a window of vertical component}

\item{visual_deci_flag}{The flag indicates if decimation is applied.
Note it will be very slow if the input data is large but without decimation}

\item{visual_deci}{A decimation factor to speed up plotting. The larger it is, the faster the plot is generated.
If visual_deci is NA, then the code automatically determines how much to decimate.}
}
\value{
The lefted indices of windows after time-domain rejection
}
\description{
This function plots each window in time domain and let users to manually select windows for rejection
}
