% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvsr_win_calc.R
\name{hvsr_win_calc}
\alias{hvsr_win_calc}
\title{HVSR calculation for the given window}
\usage{
hvsr_win_calc(
  i_win,
  h1_wins,
  h2_wins,
  v_wins,
  dt,
  rotd50_flag = TRUE,
  freq_hv_mean,
  polar_curves_flag = TRUE,
  freq_polar,
  deg_increment = 10
)
}
\arguments{
\item{i_win}{The index of running window}

\item{h1_wins}{A list of 1st horizontal component time series. Each element is for one window}

\item{h2_wins}{A list of 2nd horizontal component time series. Each element is for one window}

\item{v_wins}{A list of vertical horizontal component time series. Each element is for one window}

\item{dt}{The time step}

\item{rotd50_flag}{The flag indicates if RotD50 HVSR is calculated}

\item{freq_hv_mean}{The target frequencys of RotD50 HVSR}

\item{polar_curves_flag}{The flag indicates if polar curves are calculated}

\item{freq_polar}{The target frequencys of polar HVSR}

\item{deg_increment}{The degree increment for HVSR polar curves}
}
\value{
The HVSR for the given window
}
\description{
This function calculates RotD50 HVSR and polar HVSR for the given window
}
